+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
 ## [6.1.0-rc.1] - 2026-08-26
 
 ### Added
@@ -810,6 +812,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[Unreleased]: https://https://github.com/internetguru/laravel-model-browser/compare/staging...dev
 [6.1.0-rc.1]: https://github.com/internetguru/laravel-model-browser/releases/tag/v6.0.1
 [6.0.1]: https://https://github.com/internetguru/laravel-model-browser/compare/v6.0.0...v6.0.1
 [6.0.0]: https://https://github.com/internetguru/laravel-model-browser/compare/v5.3.3...v6.0.0